Refactor/kafka/notification v1

See merge request s11-fintech-finance-sub1/S11P21A403!224

+++ backend/picture-service/src/main/java/com/ssamba/petsi/picture_service/global/dto/NotificationProducerDto.java
@@ -11,7 +11,7 @@ public class NotificationProducerDto {
 	private String content;
 	private Long id;
 
-	public static NotificationProducerDto toNoticeProducerDto(Long pictureCnt, Long userId) {
+	public static NotificationProducerDto toNoticeProducerDto(int pictureCnt, Long userId) {
 		return new NotificationProducerDto(
 			userId,
 			"streak",
